requirements.txt
dotenv==0.9.9
et_xmlfile==2.0.0
numpy==2.2.6
opencv-python==4.12.0.88
openpyxl==3.1.5
packaging==25.0
pandas==2.3.2
pillow==11.3.0
psycopg2-binary==2.9.10
pyaes==1.6.1
pyasn1==0.6.1
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
rsa==4.9.1
six==1.17.0
Telethon==1.40.0
tzdata==2025.2
